requirements.txt
streamlit==1.36.0
pandas==2.2.2
openpyxl==3.1.5
pillow==10.3.0
